requirements.txt
aiohttp==3.7.4.post0
openpyxl==3.0.7
requests==2.26.0
tencentcloud-sdk-python==3.0.462
